requirements.txt
requests
pyslack
